refactor: organize folder structure of core-modules
