+++ website/management/commands/check_security_txt_simple.py
@@ -5,8 +5,8 @@ import requests
 from django.core.management.base import BaseCommand
 from django.utils import timezone
 
+from website.core.utils import check_security_txt
 from website.models import Domain
-from website.utils import check_security_txt
 
 logger = logging.getLogger(__name__)
