+++ website/views/organization.py
@@ -40,6 +40,13 @@ from django.views.generic.edit import CreateView
 from rest_framework import status
 from rest_framework.authtoken.models import Token
 
+from website.core.utils import (
+    format_timedelta,
+    get_client_ip,
+    get_github_issue_title,
+    rebuild_safe_url,
+    validate_file_type,
+)
 from website.forms import CaptchaForm, HuntForm, IpReportForm, RoomForm, UserProfileForm
 from website.models import (
     IP,
@@ -65,7 +72,6 @@ from website.models import (
     Winner,
 )
 from website.services.blue_sky_service import BlueSkyService
-from website.utils import format_timedelta, get_client_ip, get_github_issue_title, rebuild_safe_url, validate_file_type
 
 logger = logging.getLogger(__name__)
